Add support for validating all the problem dmp files I've collected via a script instead of manually to make sure that they still pass validation.
Add a test admin check into the RunAllBuildHeapDumpTests.ps1
Change the admin check to only be required if we need to generate the dump files (i.e. not needed for the check only mode).
Add support for the older one dmp file validation to support the problem dump files that are not created in the newer set style.

+++ RunAllBuildHeapDumpTests.ps1
@@ -8,6 +8,22 @@ Param
     [switch] $Verbose
 )
 
+Function Test-Admin {
+    $currentUser = New-Object Security.Principal.WindowsPrincipal $([Security.Principal.WindowsIdentity]::GetCurrent())
+    $currentUser.IsInRole([Security.Principal.WindowsBuiltinRole]::Administrator)
+}
+
+if($Verbose)
+{
+    $VerbosePreference = "continue"
+}
+
+if (!$CheckOnly -and (Test-Admin) -eq $false) 
+{
+    Write-Host "This script can only be run with Administrator"
+    exit
+}
+
 if(!$LastReport)
 {
     Remove-Item $ResultFile -ErrorAction:SilentlyContinue | Out-Null
